requirements.txt
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
bcrypt==3.2.0
black==21.11b1
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.0
h11==0.12.0
idna==3.3
motor==2.3.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pymongo==3.12.2
pyodbc==4.0.32
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
regex==2021.11.10
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tomli==1.2.2
typing_extensions==4.0.1
uuid==1.30
uvicorn==0.15.0
